// src/features/headless-checkout/web-components/card-number/card-type.enum.ts
export enum CardType {
  AMERICAN_EXPRESS = 'american-express',
  CHINA_UNION_PAY = 'china-union-pay',
  DEFAULT_CARD = 'default-card',
  DISCOVER = 'discover',
  HIPERCARD = 'hipercard',
  MAESTRO = 'maestro',
  MIR = 'mir',
  POSTEPAY = 'postepay',
  AURA = 'aura',
  DANKORT = 'dankort',
  DINERS_CLUBS = 'diners-clubs',
  ELO = 'elo',
  JCB = 'jcb',
  MASTERCARD = 'mastercard',
  NARANJA = 'naranja',
  VISA = 'visa',
}